remove -r to prevent opportunity to repeat command
With -r, if you press <leader><command> it has a small timout to allow an additional command to be entered.
This could be helpful if you want to press leader once and then navigate around your panes.
This is not useful with vim because when I swtich panes I often want to immediately do a vim action like go up or down.