lazyvim: switch <leader>fE and <leader>fe as well as <leader>e and <leader>E
I want the default behavior when I open the explorer to open to CWD. this is the directory I opened neovim in.
the "root" is like the project root. so if I'm inside a subprojected in a monorepo is will open that instead.

I still want this functionality, but more often then not I want to open the CWD